requirements.txt
groq
manim
matplotlib
numpy
pydantic>=2.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
langchain>=0.2.0
langchain-google-genai
jsonschema>=4.17.3
tenacity>=8.2.2

openai
litellm==1.40.3
dspy==2.5.1
regex

langchain-core==0.2.4
langchain-community==0.2.4

opencv-python
